pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristian Ticu",email = "cristian.ticu@eggnita.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.7,<0.116.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "psycopg-pool (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "boto3[crt] (>=1.37.13,<2.0.0)",
    "types-boto3[essential] (>=1.37.13,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
